@@ -40,7 +40,7 @@ type GUIConfiguration struct {
 	SendBasicAuthPrompt       bool     `protobuf:"varint,14,opt,name=send_basic_auth_prompt,json=sendBasicAuthPrompt,proto3" json:"sendBasicAuthPrompt" xml:"sendBasicAuthPrompt,attr"`
 	WebauthnUserId            []byte   `protobuf:"bytes,15,opt,name=webauthn_user_id,json=webauthnUserId,proto3" json:"webauthnUserId" xml:"webauthnUserId"`
 	WebauthnRpId              string   `protobuf:"bytes,16,opt,name=webauthn_rp_id,json=webauthnRpId,proto3" json:"webauthnRpId" xml:"webauthnRpId" default:"localhost"`
-	WebauthnOrigin            string   `protobuf:"bytes,17,opt,name=webauthn_origin,json=webauthnOrigin,proto3" json:"webauthnOrigin" xml:"webauthnOrigin" default:"https://localhost:8384"`
+	RawWebauthnOrigins        []string `protobuf:"bytes,17,rep,name=webauthn_origins,json=webauthnOrigins,proto3" json:"webauthnOrigins" xml:"webauthnOrigin"`
 }
 
 func (m *GUIConfiguration) Reset()         { *m = GUIConfiguration{} }
@@ -83,72 +83,71 @@ func init() {
 func init() { proto.RegisterFile("lib/config/guiconfiguration.proto", fileDescriptor_2a9586d611855d64) }
 
 var fileDescriptor_2a9586d611855d64 = []byte{
-	// 1029 bytes of a gzipped FileDescriptorProto
+	// 1024 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x8c, 0x56, 0xcf, 0x6f, 0xdb, 0x36,
-	0x14, 0x8e, 0xb6, 0x34, 0x89, 0xb9, 0xcc, 0xf5, 0xb4, 0x35, 0x53, 0x8b, 0xd5, 0x4c, 0x5d, 0xad,
-	0x48, 0x81, 0xc2, 0x69, 0xd3, 0x15, 0x2d, 0x82, 0x61, 0x40, 0x52, 0xa0, 0x6b, 0x90, 0x0c, 0x0b,
-	0x98, 0x05, 0x03, 0x7a, 0x11, 0x64, 0x89, 0xb1, 0x09, 0xeb, 0xd7, 0x44, 0x0a, 0x8e, 0x0f, 0xdd,
-	0x1f, 0xb0, 0xd3, 0x90, 0x9d, 0x07, 0xf4, 0x6f, 0xd8, 0x65, 0xff, 0x42, 0x6f, 0xf6, 0x69, 0xd8,
-	0x89, 0x40, 0x9d, 0x9b, 0x2e, 0x03, 0x74, 0xcc, 0x69, 0x20, 0x65, 0xc9, 0x92, 0xa3, 0xac, 0xbb,
-	0x91, 0xdf, 0xf7, 0xf1, 0x7d, 0xef, 0x91, 0xef, 0xd9, 0x02, 0x77, 0x1c, 0xd2, 0xd9, 0xb4, 0x7c,
-	0xef, 0x84, 0x74, 0x37, 0xbb, 0x11, 0x49, 0x57, 0x51, 0x68, 0x32, 0xe2, 0x7b, 0xed, 0x20, 0xf4,
-	0x99, 0xaf, 0x2e, 0xa5, 0xe0, 0xad, 0x9b, 0x05, 0xa9, 0x19, 0xb1, 0x9e, 0xeb, 0xdb, 0x38, 0x95,
-	0xdc, 0xaa, 0xe1, 0x53, 0x96, 0x2e, 0x5b, 0xff, 0x34, 0x40, 0xe3, 0xdb, 0xe3, 0xbd, 0xe7, 0xc5,
-	0x40, 0x6a, 0x07, 0x2c, 0x63, 0xcf, 0xec, 0x38, 0xd8, 0xd6, 0x94, 0x75, 0x65, 0x63, 0x65, 0xf7,
-	0x65, 0xcc, 0x61, 0x06, 0x25, 0x1c, 0xde, 0x39, 0x75, 0x9d, 0xed, 0xd6, 0x74, 0xff, 0xc0, 0x64,
-	0x2c, 0x6c, 0xad, 0xdb, 0xf8, 0xc4, 0x8c, 0x1c, 0xb6, 0xdd, 0x62, 0x61, 0x84, 0x5b, 0xf1, 0x48,
-	0x5f, 0x2d, 0xf2, 0x17, 0x23, 0x7d, 0x51, 0x10, 0x28, 0x8b, 0xa2, 0xbe, 0x06, 0xcb, 0xa6, 0x6d,
-	0x87, 0x98, 0x52, 0xed, 0x83, 0x75, 0x65, 0xa3, 0xb6, 0x6b, 0x4d, 0x38, 0x04, 0xc8, 0x1c, 0xec,
-	0xa4, 0xa8, 0x70, 0x9c, 0x0a, 0x12, 0x0e, 0xef, 0x49, 0xc7, 0xe9, 0xbe, 0x60, 0xf6, 0x68, 0xeb,
-	0x69, 0xfb, 0x61, 0xfb, 0x61, 0xfb, 0xd1, 0xf6, 0xb3, 0xc7, 0xcf, 0xbe, 0x6a, 0x5d, 0x8c, 0xf4,
-	0x7a, 0x19, 0x3a, 0x1b, 0xeb, 0x85, 0xa0, 0x28, 0x0b, 0xa9, 0xfe, 0xa5, 0x80, 0xcf, 0x23, 0x8f,
-	0x9c, 0x1a, 0xd4, 0xb7, 0xfa, 0x98, 0x19, 0x01, 0x0e, 0x5d, 0x42, 0x29, 0xf1, 0x3d, 0xaa, 0x7d,
-	0x28, 0xf3, 0xf9, 0x5d, 0x99, 0x70, 0xa8, 0x21, 0x73, 0x70, 0xec, 0x91, 0xd3, 0x23, 0xa9, 0x3a,
-	0x9c, 0x89, 0x62, 0x0e, 0x6f, 0x44, 0x55, 0x44, 0xc2, 0xe1, 0x97, 0x32, 0xd9, 0x4a, 0xf6, 0x81,
-	0xef, 0x12, 0x86, 0xdd, 0x80, 0x0d, 0xc5, 0x15, 0xc1, 0xf7, 0x68, 0xce, 0xc6, 0xfa, 0x95, 0x09,
-	0xa0, 0x6a, 0x7b, 0xf5, 0x05, 0x58, 0x8c, 0x28, 0x0e, 0xb5, 0x45, 0x59, 0xc4, 0x56, 0xcc, 0xa1,
-	0xdc, 0x27, 0x1c, 0x7e, 0x96, 0xa6, 0x45, 0x71, 0x58, 0xce, 0xa2, 0x5e, 0x86, 0x90, 0xd4, 0xab,
-	0xaf, 0xc0, 0x4a, 0x60, 0x52, 0x3a, 0xf0, 0x43, 0x5b, 0xbb, 0x26, 0x63, 0x7d, 0x13, 0x73, 0x98,
-	0x63, 0x09, 0x87, 0x9a, 0x8c, 0x97, 0x01, 0xe5, 0x98, 0xea, 0x65, 0x18, 0xe5, 0x67, 0x55, 0x17,
-	0xd4, 0x44, 0x47, 0x1a, 0xa2, 0x25, 0xb5, 0xa5, 0x75, 0x65, 0xa3, 0xbe, 0xd5, 0x68, 0xa7, 0xad,
-	0xda, 0xde, 0x89, 0x58, 0xef, 0x3b, 0xdf, 0xc6, 0xa9, 0x9d, 0x39, 0xdd, 0xe5, 0x76, 0x19, 0x30,
-	0x67, 0x77, 0x19, 0x46, 0xf9, 0x59, 0x15, 0x83, 0xe5, 0x88, 0x62, 0x83, 0x39, 0x54, 0x5b, 0x96,
-	0xed, 0x7c, 0x30, 0xe1, 0xb0, 0x26, 0x2e, 0x96, 0xe2, 0x1f, 0x0e, 0x8e, 0x62, 0x0e, 0x97, 0x22,
-	0xb9, 0x4a, 0x38, 0xac, 0x4b, 0x17, 0xe6, 0xd0, 0xb4, 0xad, 0xe3, 0x91, 0xbe, 0x92, 0x6d, 0x92,
-	0x91, 0x3e, 0xd5, 0x9d, 0x8d, 0xf5, 0xd9, 0x71, 0x24, 0x41, 0x87, 0x0a, 0x1b, 0x33, 0x20, 0x46,
-	0x1f, 0x0f, 0xb5, 0x15, 0x79, 0x61, 0xc2, 0x66, 0x69, 0xe7, 0x70, 0x6f, 0x1f, 0x0f, 0x85, 0x87,
-	0x19, 0x90, 0x7d, 0x3c, 0x4c, 0x38, 0x5c, 0x4b, 0x2b, 0x09, 0x48, 0x1f, 0x0f, 0xcb, 0x75, 0x34,
-	0xe6, 0xc1, 0xb3, 0xb1, 0x3e, 0x8d, 0x80, 0xa6, 0xe7, 0xd5, 0xdf, 0x14, 0x70, 0x83, 0x78, 0x14,
-	0x5b, 0x51, 0x88, 0x0d, 0xd3, 0x76, 0x89, 0x67, 0x98, 0x96, 0x25, 0xe6, 0xa8, 0x26, 0x8b, 0x33,
-	0x62, 0x0e, 0x3f, 0xcd, 0x04, 0x3b, 0x82, 0xdf, 0x91, 0x74, 0xc2, 0xe1, 0x5d, 0x69, 0x5c, 0xc1,
-	0x95, 0xb3, 0xb8, 0xfd, 0x9f, 0x0a, 0x54, 0x15, 0x5c, 0xdd, 0x07, 0xd7, 0x58, 0x0f, 0xbb, 0x58,
-	0x03, 0xb2, 0xf4, 0x27, 0x31, 0x87, 0x29, 0x90, 0x70, 0x78, 0x3b, 0xbd, 0x53, 0xb1, 0x2b, 0x8c,
-	0xee, 0x74, 0x21, 0x66, 0x76, 0x79, 0xba, 0x46, 0xe9, 0x11, 0xf5, 0x18, 0xd4, 0x6c, 0xdc, 0x89,
-	0xba, 0x5d, 0xe2, 0x75, 0xb5, 0x8f, 0x64, 0x55, 0x4f, 0x63, 0x0e, 0x67, 0x60, 0xde, 0xcd, 0x39,
-	0x92, 0x3f, 0x57, 0xbd, 0x0c, 0xa1, 0xd9, 0x21, 0xf5, 0x4f, 0x05, 0x68, 0xf9, 0xcd, 0xd1, 0x3e,
-	0x09, 0x8c, 0x9e, 0x4f, 0x99, 0x61, 0xf5, 0xb0, 0xd5, 0xd7, 0x56, 0xa5, 0xcd, 0xcf, 0x62, 0xae,
-	0x33, 0xcd, 0x51, 0x9f, 0x04, 0x2f, 0x7d, 0xca, 0xa4, 0x20, 0x9f, 0xeb, 0x4a, 0x76, 0x6e, 0xae,
-	0xdf, 0xa3, 0x49, 0x46, 0x7a, 0xb5, 0x09, 0xba, 0x04, 0x3f, 0x17, 0xb0, 0xfa, 0x87, 0x02, 0xbe,
-	0x98, 0xbd, 0xb9, 0xe3, 0xf8, 0x03, 0xe3, 0x24, 0x34, 0x5d, 0x6c, 0x38, 0xbe, 0x69, 0x8b, 0x4b,
-	0xfa, 0x58, 0x66, 0xff, 0x53, 0xcc, 0xe1, 0xcd, 0xfc, 0x75, 0x84, 0xec, 0x85, 0x50, 0x1d, 0xa4,
-	0xa2, 0x84, 0xc3, 0xfb, 0xe5, 0x06, 0x98, 0x57, 0x94, 0xab, 0xb8, 0xfb, 0x3f, 0x74, 0xe8, 0x6a,
-	0x3b, 0xf5, 0x17, 0x05, 0xac, 0x51, 0xec, 0xd9, 0x46, 0xc7, 0xa4, 0xc4, 0x32, 0xe4, 0xc4, 0x07,
-	0xa1, 0xef, 0x06, 0x4c, 0xab, 0xcb, 0x74, 0x8f, 0x45, 0xa7, 0x0a, 0xc5, 0xae, 0x10, 0x88, 0xc1,
-	0x3f, 0x94, 0x74, 0xc2, 0x61, 0x53, 0x26, 0x5a, 0xc1, 0xe5, 0xef, 0xac, 0x5d, 0x45, 0xa2, 0xaa,
-	0x90, 0xea, 0x09, 0x68, 0x0c, 0x70, 0x47, 0xe4, 0xe0, 0x19, 0xe2, 0xf7, 0xcd, 0x20, 0xb6, 0x76,
-	0x7d, 0x5d, 0xd9, 0x58, 0xdd, 0xfd, 0x3a, 0xe6, 0xb0, 0x9e, 0x71, 0xc7, 0x14, 0x87, 0x7b, 0x76,
-	0xde, 0x5e, 0x65, 0x58, 0xb6, 0x57, 0x19, 0x42, 0x73, 0x7b, 0xf5, 0x35, 0xc8, 0x11, 0x23, 0x0c,
-	0x84, 0x4b, 0x43, 0x0e, 0xc4, 0x8f, 0x31, 0x87, 0xab, 0x19, 0x83, 0x02, 0xe9, 0x71, 0xaf, 0xe4,
-	0x21, 0xc0, 0xc2, 0x78, 0x38, 0xbe, 0x65, 0x3a, 0xa2, 0x3f, 0xe5, 0x7f, 0x69, 0x51, 0x74, 0x31,
-	0xd2, 0x6b, 0x39, 0x8b, 0x4a, 0x94, 0xfa, 0x46, 0x01, 0xd7, 0x73, 0x7f, 0x3f, 0x24, 0x5d, 0xe2,
-	0x69, 0x9f, 0xc8, 0x04, 0x06, 0xc5, 0x32, 0xbf, 0x97, 0x4c, 0xc2, 0xe1, 0x93, 0x52, 0x0a, 0x29,
-	0x5c, 0x48, 0xa2, 0xc7, 0x58, 0x40, 0xb7, 0x37, 0x37, 0x73, 0xbb, 0xf4, 0x6f, 0xb6, 0x78, 0x0f,
-	0xe9, 0x99, 0x8b, 0x91, 0xbe, 0x56, 0x2d, 0x46, 0x73, 0xca, 0xdd, 0xfd, 0xb7, 0xef, 0x9a, 0x0b,
-	0xe3, 0x77, 0xcd, 0x85, 0xb7, 0x93, 0xa6, 0x32, 0x9e, 0x34, 0x95, 0x5f, 0xcf, 0x9b, 0x0b, 0x6f,
-	0xce, 0x9b, 0xca, 0xf8, 0xbc, 0xb9, 0xf0, 0xf7, 0x79, 0x73, 0xe1, 0xd5, 0xfd, 0x2e, 0x61, 0xbd,
-	0xa8, 0xd3, 0xb6, 0x7c, 0x77, 0x93, 0x0e, 0x3d, 0x8b, 0xf5, 0x88, 0xd7, 0x2d, 0xac, 0x66, 0x1f,
-	0x36, 0x9d, 0x25, 0xf9, 0x15, 0xf3, 0xf8, 0xdf, 0x00, 0x00, 0x00, 0xff, 0xff, 0x83, 0x2f, 0x37,
-	0xc7, 0x18, 0x09, 0x00, 0x00,
+	0x14, 0x8e, 0xd6, 0x36, 0x89, 0xb5, 0xcc, 0xf5, 0xb4, 0xb5, 0x53, 0x8b, 0xd5, 0x74, 0x5d, 0xad,
+	0x48, 0x81, 0xc2, 0x69, 0xd3, 0x0d, 0x2d, 0x82, 0x61, 0x40, 0x5c, 0xa0, 0x6b, 0x90, 0x0c, 0x0b,
+	0x98, 0x05, 0x05, 0x7a, 0x11, 0x64, 0x89, 0xb1, 0x09, 0xeb, 0xd7, 0x44, 0x0a, 0x8e, 0x0f, 0xdd,
+	0x1f, 0xb0, 0xd3, 0x90, 0xed, 0x3a, 0x60, 0x7f, 0xc3, 0x2e, 0xfb, 0x17, 0x7a, 0xb3, 0x4f, 0xc3,
+	0x4e, 0x04, 0x6a, 0xdf, 0x74, 0xd4, 0x31, 0xa7, 0x81, 0xd4, 0x0f, 0x4b, 0x8e, 0xb2, 0xee, 0x46,
+	0x7e, 0xdf, 0xc7, 0xf7, 0xbd, 0x47, 0xbe, 0x67, 0x4b, 0xbe, 0x6b, 0xe3, 0xde, 0x96, 0xe9, 0xb9,
+	0x27, 0xb8, 0xbf, 0xd5, 0x0f, 0x71, 0xb2, 0x0a, 0x03, 0x83, 0x62, 0xcf, 0xed, 0xf8, 0x81, 0x47,
+	0x3d, 0x65, 0x35, 0x01, 0x6f, 0xdf, 0x2a, 0x48, 0x8d, 0x90, 0x0e, 0x1c, 0xcf, 0x42, 0x89, 0xe4,
+	0x76, 0x0d, 0x9d, 0xd2, 0x64, 0xd9, 0x9e, 0x37, 0xe4, 0xc6, 0xb7, 0xc7, 0x7b, 0xcf, 0x8b, 0x81,
+	0x94, 0x9e, 0xbc, 0x86, 0x5c, 0xa3, 0x67, 0x23, 0x4b, 0x95, 0x5a, 0xd2, 0xe6, 0x7a, 0xf7, 0x65,
+	0xc4, 0x40, 0x06, 0xc5, 0x0c, 0xdc, 0x3d, 0x75, 0xec, 0x9d, 0x76, 0xba, 0x7f, 0x68, 0x50, 0x1a,
+	0xb4, 0x5b, 0x16, 0x3a, 0x31, 0x42, 0x9b, 0xee, 0xb4, 0x69, 0x10, 0xa2, 0x76, 0x34, 0xd1, 0x36,
+	0x8a, 0xfc, 0xf9, 0x44, 0xbb, 0xca, 0x09, 0x98, 0x45, 0x51, 0xde, 0xc8, 0x6b, 0x86, 0x65, 0x05,
+	0x88, 0x10, 0xf5, 0x83, 0x96, 0xb4, 0x59, 0xeb, 0x9a, 0x33, 0x06, 0x64, 0x68, 0x8c, 0x76, 0x13,
+	0x94, 0x3b, 0xa6, 0x82, 0x98, 0x81, 0xfb, 0xc2, 0x31, 0xdd, 0x17, 0xcc, 0x1e, 0x6f, 0x3f, 0xed,
+	0x3c, 0xea, 0x3c, 0xea, 0x3c, 0xde, 0x79, 0xf6, 0xe4, 0xd9, 0x97, 0xed, 0xf3, 0x89, 0x56, 0x2f,
+	0x43, 0x67, 0x53, 0xad, 0x10, 0x14, 0x66, 0x21, 0x95, 0xbf, 0x25, 0xf9, 0xb3, 0xd0, 0xc5, 0xa7,
+	0x3a, 0xf1, 0xcc, 0x21, 0xa2, 0xba, 0x8f, 0x02, 0x07, 0x13, 0x82, 0x3d, 0x97, 0xa8, 0x57, 0x44,
+	0x3e, 0xbf, 0x4b, 0x33, 0x06, 0x54, 0x68, 0x8c, 0x8e, 0x5d, 0x7c, 0x7a, 0x24, 0x54, 0x87, 0x0b,
+	0x51, 0xc4, 0xc0, 0x8d, 0xb0, 0x8a, 0x88, 0x19, 0xf8, 0x42, 0x24, 0x5b, 0xc9, 0x3e, 0xf4, 0x1c,
+	0x4c, 0x91, 0xe3, 0xd3, 0x31, 0xbf, 0x22, 0xf0, 0x1e, 0xcd, 0xd9, 0x54, 0xbb, 0x34, 0x01, 0x58,
+	0x6d, 0xaf, 0xbc, 0x90, 0xaf, 0x86, 0x04, 0x05, 0xea, 0x55, 0x51, 0xc4, 0x76, 0xc4, 0x80, 0xd8,
+	0xc7, 0x0c, 0x7c, 0x9a, 0xa4, 0x45, 0x50, 0x50, 0xce, 0xa2, 0x5e, 0x86, 0xa0, 0xd0, 0x2b, 0xaf,
+	0xe5, 0x75, 0xdf, 0x20, 0x64, 0xe4, 0x05, 0x96, 0x7a, 0x4d, 0xc4, 0xfa, 0x26, 0x62, 0x20, 0xc7,
+	0x62, 0x06, 0x54, 0x11, 0x2f, 0x03, 0xca, 0x31, 0x95, 0x8b, 0x30, 0xcc, 0xcf, 0x2a, 0x8e, 0x5c,
+	0xe3, 0x1d, 0xa9, 0xf3, 0x96, 0x54, 0x57, 0x5b, 0xd2, 0x66, 0x7d, 0xbb, 0xd1, 0x49, 0x5a, 0xb5,
+	0xb3, 0x1b, 0xd2, 0xc1, 0x77, 0x9e, 0x85, 0x12, 0x3b, 0x23, 0xdd, 0xe5, 0x76, 0x19, 0xb0, 0x64,
+	0x77, 0x11, 0x86, 0xf9, 0x59, 0x05, 0xc9, 0x6b, 0x21, 0x41, 0x3a, 0xb5, 0x89, 0xba, 0x26, 0xda,
+	0xf9, 0x60, 0xc6, 0x40, 0x8d, 0x5f, 0x2c, 0x41, 0x3f, 0x1c, 0x1c, 0x45, 0x0c, 0xac, 0x86, 0x62,
+	0x15, 0x33, 0x50, 0x17, 0x2e, 0xd4, 0x26, 0x49, 0x5b, 0x47, 0x13, 0x6d, 0x3d, 0xdb, 0xc4, 0x13,
+	0x2d, 0xd5, 0x9d, 0x4d, 0xb5, 0xc5, 0x71, 0x28, 0x40, 0x9b, 0x70, 0x1b, 0xc3, 0xc7, 0xfa, 0x10,
+	0x8d, 0xd5, 0x75, 0x71, 0x61, 0xdc, 0x66, 0x75, 0xf7, 0x70, 0x6f, 0x1f, 0x8d, 0xb9, 0x87, 0xe1,
+	0xe3, 0x7d, 0x34, 0x8e, 0x19, 0xb8, 0x99, 0x54, 0xe2, 0xe3, 0x21, 0x1a, 0x97, 0xeb, 0x68, 0x2c,
+	0x83, 0x67, 0x53, 0x2d, 0x8d, 0x00, 0xd3, 0xf3, 0xca, 0xaf, 0x92, 0x7c, 0x03, 0xbb, 0x04, 0x99,
+	0x61, 0x80, 0x74, 0xc3, 0x72, 0xb0, 0xab, 0x1b, 0xa6, 0xc9, 0xe7, 0xa8, 0x26, 0x8a, 0xd3, 0x23,
+	0x06, 0x3e, 0xc9, 0x04, 0xbb, 0x9c, 0xdf, 0x15, 0x74, 0xcc, 0xc0, 0x3d, 0x61, 0x5c, 0xc1, 0x95,
+	0xb3, 0xb8, 0xf3, 0x9f, 0x0a, 0x58, 0x15, 0x5c, 0xd9, 0x97, 0xaf, 0xd1, 0x01, 0x72, 0x90, 0x2a,
+	0x8b, 0xd2, 0xbf, 0x8a, 0x18, 0x48, 0x80, 0x98, 0x81, 0x3b, 0xc9, 0x9d, 0xf2, 0x5d, 0x61, 0x74,
+	0xd3, 0x05, 0x9f, 0xd9, 0xb5, 0x74, 0x0d, 0x93, 0x23, 0xca, 0xb1, 0x5c, 0xb3, 0x50, 0x2f, 0xec,
+	0xf7, 0xb1, 0xdb, 0x57, 0x3f, 0x14, 0x55, 0x3d, 0x8d, 0x18, 0x58, 0x80, 0x79, 0x37, 0xe7, 0x48,
+	0xfe, 0x5c, 0xf5, 0x32, 0x04, 0x17, 0x87, 0x94, 0xbf, 0x24, 0x59, 0xcd, 0x6f, 0x8e, 0x0c, 0xb1,
+	0xaf, 0x0f, 0x3c, 0x42, 0x75, 0x73, 0x80, 0xcc, 0xa1, 0xba, 0x21, 0x6c, 0x7e, 0xe2, 0x73, 0x9d,
+	0x69, 0x8e, 0x86, 0xd8, 0x7f, 0xe9, 0x11, 0x2a, 0x04, 0xf9, 0x5c, 0x57, 0xb2, 0x4b, 0x73, 0xfd,
+	0x1e, 0x4d, 0x3c, 0xd1, 0xaa, 0x4d, 0xe0, 0x05, 0xf8, 0x39, 0x87, 0x95, 0x3f, 0x25, 0xf9, 0xf3,
+	0xc5, 0x9b, 0xdb, 0xb6, 0x37, 0xd2, 0x4f, 0x02, 0xc3, 0x41, 0xba, 0xed, 0x19, 0x16, 0xbf, 0xa4,
+	0x8f, 0x44, 0xf6, 0x3f, 0x46, 0x0c, 0xdc, 0xca, 0x5f, 0x87, 0xcb, 0x5e, 0x70, 0xd5, 0x41, 0x22,
+	0x8a, 0x19, 0x78, 0x50, 0x6e, 0x80, 0x65, 0x45, 0xb9, 0x8a, 0x7b, 0xff, 0x43, 0x07, 0x2f, 0xb7,
+	0x53, 0x7e, 0x96, 0xe4, 0x9b, 0x04, 0xb9, 0x96, 0xde, 0x33, 0x08, 0x36, 0x75, 0x31, 0xf1, 0x7e,
+	0xe0, 0x39, 0x3e, 0x55, 0xeb, 0x22, 0xdd, 0x63, 0xde, 0xa9, 0x5c, 0xd1, 0xe5, 0x02, 0x3e, 0xf8,
+	0x87, 0x82, 0x8e, 0x19, 0x68, 0x8a, 0x44, 0x2b, 0xb8, 0xfc, 0x9d, 0xd5, 0xcb, 0x48, 0x58, 0x15,
+	0x52, 0x39, 0x91, 0x1b, 0x23, 0xd4, 0xe3, 0x39, 0xb8, 0x3a, 0xff, 0x7d, 0xd3, 0xb1, 0xa5, 0x5e,
+	0x6f, 0x49, 0x9b, 0x1b, 0xdd, 0xaf, 0x23, 0x06, 0xea, 0x19, 0x77, 0x4c, 0x50, 0xb0, 0x67, 0xe5,
+	0xed, 0x55, 0x86, 0x45, 0x7b, 0x95, 0x21, 0xb8, 0xb4, 0x57, 0xde, 0xc8, 0x39, 0xa2, 0x07, 0x3e,
+	0x77, 0x69, 0x88, 0x81, 0x78, 0x15, 0x31, 0xb0, 0x91, 0x31, 0xd0, 0x17, 0x1e, 0xf7, 0x4b, 0x1e,
+	0x1c, 0x2c, 0x8c, 0x87, 0xed, 0x99, 0x86, 0xcd, 0xfb, 0x53, 0xfc, 0x97, 0x16, 0x45, 0xe7, 0x13,
+	0xad, 0x96, 0xb3, 0xb0, 0x44, 0x29, 0xbf, 0x49, 0x85, 0x3a, 0xbd, 0x00, 0xf7, 0xb1, 0x4b, 0xd4,
+	0x8f, 0x5b, 0x57, 0x36, 0x6b, 0x5d, 0x3c, 0x63, 0x40, 0x81, 0xc6, 0xe8, 0x55, 0x4a, 0x7f, 0x9f,
+	0xb0, 0x11, 0x03, 0xd7, 0x47, 0x65, 0xe8, 0x42, 0xf9, 0x09, 0x5e, 0x2a, 0x3f, 0x81, 0xce, 0xa6,
+	0x5a, 0x45, 0x48, 0xb8, 0x1c, 0xb0, 0xbb, 0xff, 0xf6, 0x5d, 0x73, 0x65, 0xfa, 0xae, 0xb9, 0xf2,
+	0x76, 0xd6, 0x94, 0xa6, 0xb3, 0xa6, 0xf4, 0xcb, 0xbc, 0xb9, 0xf2, 0xc7, 0xbc, 0x29, 0x4d, 0xe7,
+	0xcd, 0x95, 0x7f, 0xe6, 0xcd, 0x95, 0xd7, 0x0f, 0xfa, 0x98, 0x0e, 0xc2, 0x5e, 0xc7, 0xf4, 0x9c,
+	0x2d, 0x32, 0x76, 0x4d, 0x3a, 0xc0, 0x6e, 0xbf, 0xb0, 0x5a, 0x7c, 0xcc, 0xf4, 0x56, 0xc5, 0x97,
+	0xcb, 0x93, 0x7f, 0x03, 0x00, 0x00, 0xff, 0xff, 0xeb, 0x55, 0x01, 0x3a, 0x0c, 0x09, 0x00, 0x00,
 }
 
 func (m *GUIConfiguration) Marshal() (dAtA []byte, err error) {
@@ -171,14 +170,16 @@ func (m *GUIConfiguration) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.WebauthnOrigin) > 0 {
-		i -= len(m.WebauthnOrigin)
-		copy(dAtA[i:], m.WebauthnOrigin)
-		i = encodeVarintGuiconfiguration(dAtA, i, uint64(len(m.WebauthnOrigin)))
-		i--
-		dAtA[i] = 0x1
-		i--
-		dAtA[i] = 0x8a
+	if len(m.RawWebauthnOrigins) > 0 {
+		for iNdEx := len(m.RawWebauthnOrigins) - 1; iNdEx >= 0; iNdEx-- {
+			i -= len(m.RawWebauthnOrigins[iNdEx])
+			copy(dAtA[i:], m.RawWebauthnOrigins[iNdEx])
+			i = encodeVarintGuiconfiguration(dAtA, i, uint64(len(m.RawWebauthnOrigins[iNdEx])))
+			i--
+			dAtA[i] = 0x1
+			i--
+			dAtA[i] = 0x8a
+		}
 	}
 	if len(m.WebauthnRpId) > 0 {
 		i -= len(m.WebauthnRpId)
@@ -389,9 +390,11 @@ func (m *GUIConfiguration) ProtoSize() (n int) {
 	if l > 0 {
 		n += 2 + l + sovGuiconfiguration(uint64(l))
 	}
-	l = len(m.WebauthnOrigin)
-	if l > 0 {
-		n += 2 + l + sovGuiconfiguration(uint64(l))
+	if len(m.RawWebauthnOrigins) > 0 {
+		for _, s := range m.RawWebauthnOrigins {
+			l = len(s)
+			n += 2 + l + sovGuiconfiguration(uint64(l))
+		}
 	}
 	return n
 }
@@ -850,7 +853,7 @@ func (m *GUIConfiguration) Unmarshal(dAtA []byte) error {
 			iNdEx = postIndex
 		case 17:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field WebauthnOrigin", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field RawWebauthnOrigins", wireType)
 			}
 			var stringLen uint64
 			for shift := uint(0); ; shift += 7 {
@@ -878,7 +881,7 @@ func (m *GUIConfiguration) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.WebauthnOrigin = string(dAtA[iNdEx:postIndex])
+			m.RawWebauthnOrigins = append(m.RawWebauthnOrigins, string(dAtA[iNdEx:postIndex]))
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
