@@ -6,7 +6,11 @@
 
 package config
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+)
 
 func TestMigrateCrashReporting(t *testing.T) {
 	// When migrating from pre-crash-reporting configs, crash reporting is
@@ -31,6 +35,68 @@ func TestMigrateCrashReporting(t *testing.T) {
 		migrationsMut.Unlock()
 		if cfg.Options.CREnabled != tc.enabled {
 			t.Errorf("%d: unexpected result, CREnabled: %v != %v", i, cfg.Options.CREnabled, tc.enabled)
+		}
+	}
+}
+
+func TestMigration38(t *testing.T) {
+	{
+		cfg := Configuration{
+			Options: OptionsConfiguration{
+				UnackedNotificationIDs: []string{"foo", "authenticationUserAndPassword", "bar"},
+			},
+			GUI: GUIConfiguration{
+				RawUseTLS: false,
+			},
+		}
+		migrateToConfigV38(&cfg)
+		if !cmp.Equal(cfg.Options.UnackedNotificationIDs, []string{"foo", "guiAuthentication", "bar"}) {
+			t.Error("Expected notification \"authenticationUserAndPassword\" to be renamed to \"guiAuthentication\"")
+		}
+		if cfg.GUI.WebauthnRpId != "localhost" {
+			t.Error("Expected GUI.WebauthnRpId to be set to \"localhost\"")
+		}
+		if !cmp.Equal(cfg.GUI.RawWebauthnOrigins, []string{"https://localhost:8384", "http://localhost:8384"}) {
+			t.Error("Expected GUI.WebauthnRpId to be set to default values")
+		}
+	}
+
+	{
+		cfg := Configuration{
+			GUI: GUIConfiguration{
+				RawAddress: "127.0.0.1:8888",
+				RawUseTLS:  true,
+			},
+		}
+		migrateToConfigV38(&cfg)
+		if !cmp.Equal(cfg.GUI.RawWebauthnOrigins, []string{"https://localhost:8888"}) {
+			t.Error("Expected GUI.WebauthnRpId to be set to default values with port 8888 and HTTPS only")
+		}
+	}
+
+	{
+		cfg := Configuration{
+			GUI: GUIConfiguration{
+				RawAddress: "127.0.0.1:443",
+				RawUseTLS:  true,
+			},
+		}
+		migrateToConfigV38(&cfg)
+		if !cmp.Equal(cfg.GUI.RawWebauthnOrigins, []string{"https://localhost"}) {
+			t.Error("Expected GUI.WebauthnRpId to be set to default values with implicit HTTPS port and HTTPS only")
+		}
+	}
+
+	{
+		cfg := Configuration{
+			GUI: GUIConfiguration{
+				RawAddress: "127.0.0.1:80",
+				RawUseTLS:  false,
+			},
+		}
+		migrateToConfigV38(&cfg)
+		if !cmp.Equal(cfg.GUI.RawWebauthnOrigins, []string{"https://localhost:80", "http://localhost"}) {
+			t.Error("Expected GUI.WebauthnRpId to be set to default values with implicit HTTP port")
 		}
 	}
 }
